+++ src/components/index.ts
@@ -5,7 +5,8 @@ import { ListCard } from "./ListCard";
 import { SeriesCard } from "./SeriesCard";
 import { SearchBar } from "./SearchBar";
 import { GlobalSearch } from "./GlobalSearch";
+import { HomepageVisual } from "./HomepageVisual";
 import { Breadcrumbs } from "./Breadcrumbs";
 import { BookCardSkeleton, PersonCardSkeleton, ListCardSkeleton, EmptyState } from "./Skeletons";
 
-export { SectionHeading, BookCard, PersonCard, ListCard, SeriesCard, SearchBar, GlobalSearch, Breadcrumbs, BookCardSkeleton, PersonCardSkeleton, ListCardSkeleton, EmptyState };
+export { SectionHeading, BookCard, PersonCard, ListCard, SeriesCard, SearchBar, GlobalSearch, HomepageVisual, Breadcrumbs, BookCardSkeleton, PersonCardSkeleton, ListCardSkeleton, EmptyState };